backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
neo4j==5.28.0
anthropic>=0.40.0
openai>=1.50.0
gitpython==3.1.43
tree-sitter==0.23.0
tree-sitter-python==0.23.0
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
tree-sitter-go==0.23.0
tree-sitter-rust==0.23.0
pydantic>=2.10.0
numpy>=1.26.0
